Remove dead code and consolidate test utilities
- Remove unused extract_column() from parser.rs
- Remove unused LOOSE_TOL constant and rel_approx_eq() from test files
- Create shared test utilities module (src/tests/common.rs) with:
  - Common tolerance constants
  - approx_eq() and rel_approx_eq() helper functions
  - Test data generators (aml_maintained, aml_nonmaintained, aml_combined, etc.)
  - LungData struct and helper functions
- Refactor test files to use shared utilities
