+++ src/tests/mod.rs
@@ -1,3 +1,4 @@
+pub mod common;
 pub mod r_exact_validation;
 pub mod r_survival_validation;
 pub mod r_validation_tests;
